// orbito-frontend/src/core/api/generated/models/paymentDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Orbito API
 * OpenAPI spec version: v1
 */

export interface PaymentDto {
  id?: string;
  tenantId?: string;
  subscriptionId?: string;
  clientId?: string;
  amount?: number;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  externalTransactionId?: string | null;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  externalPaymentId?: string | null;
  /** @nullable */
  paymentMethodId?: string | null;
  createdAt?: string;
  /** @nullable */
  processedAt?: string | null;
  /** @nullable */
  failedAt?: string | null;
  /** @nullable */
  refundedAt?: string | null;
  /** @nullable */
  failureReason?: string | null;
}